py: Move fixtures to own module
